requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mysqlclient==2.2.7
propcache==0.3.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
gunicorn
